// common/data/level-type.ts
export const enum LevelType {
  ATK = "ATK",
  DEF = "DEF",
  HP = "HP",
  BLUE_KEY = "Blue key",
  GREEN_CLUE_KEY = "Green blue key",
  CRIMSON_KEY = "Crimson key",
  PLATINUM_KEY = "Platinum key",
  VIOLENT_KEY = "Violet key",
  YELLOW_KEY = "Yellow key",
}

export const LEVEL_TYPES: LevelType[] = [
  LevelType.ATK,
  LevelType.DEF,
  LevelType.HP,
  LevelType.BLUE_KEY,
  LevelType.GREEN_CLUE_KEY,
  LevelType.CRIMSON_KEY,
  LevelType.PLATINUM_KEY,
  LevelType.VIOLENT_KEY,
  LevelType.YELLOW_KEY,
]